DEV: breaking messy main.cxx into separate files to make      it easier to manage
