+++ BuildWithMake/Release/developer_wrapper_scripts/developer-simvascular-mac_osx-script
@@ -12,7 +12,7 @@ $SV_HOME/BuildWithMake/Bin/sv.app/Contents/MacOS/sv $*
 esac
 ;;
 *)
-$SV_HOME/BuildWithMake/Bin/sv.app/Contents/MacOS/sv $SV_HOME/Tcl/SimVascular_2.0/simvascular_startup.tcl -tcl -qt $*
+$SV_HOME/BuildWithMake/Bin/sv.app/Contents/MacOS/sv $SV_HOME/Tcl/SimVascular_2.0/simvascular_startup.tcl -tk -qt $*
 ;;
 esac
