+++ Code/CMake/SimVascularOptions.cmake
@@ -140,6 +140,8 @@ option(SV_USE_TETGEN_ADAPTOR "Option to use open source mesh adaption" ON)
 
 option(SV_USE_PYTHON "Use Python" OFF)
 
+option(SV_USE_TCL "Use Tcl" ON)
+
 option(SV_USE_SOLVERIO "Use SolverIO" OFF)
 #-----------------------------------------------------------------------------
